Fix building on 6.12 due to missing definition of rdtsc().
Something else must have been indirectly pulling in asm/msr.h.
Explicitly add it to the includes.

This should be fine for kernels newer than 3.x. 3.x and earlier
never had a rdtsc() function by the looks of it, so I'm not sure
they would even build today. (That could be a separate fix if
needed in the future.)

Fixes #4.